Move parquet cascading schemes to subprojects
