+++ scalding-parquet-cascading/src/test/resources/names.txt
