+++ scalding-parquet-scrooge-cascading/src/test/resources/names.txt
