--- scalding-parquet-scrooge/src/main/java/com/twitter/scalding/parquet/scrooge/ScroogeSchemaConversionException.java
@@ -1,36 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements.  See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership.  The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License.  You may obtain a copy of the License at
- *
- *   http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-package com.twitter.scalding.parquet.scrooge;
-
-import org.apache.parquet.ParquetRuntimeException;
-
-/**
- * Throw this exception when there is an error converting a Scrooge class to
- * thrift schema
- */
-class ScroogeSchemaConversionException extends ParquetRuntimeException {
-  public ScroogeSchemaConversionException(String message, Throwable cause) {
-    super(message, cause);
-  }
-
-  public ScroogeSchemaConversionException(String message) {
-    super(message);
-  }
-}
-
