+++ scalding-parquet-scrooge/src/main/scala/com/twitter/scalding/parquet/scrooge/package.scala
@@ -1,0 +1,6 @@
+package com.twitter.scalding.parquet
+
+package object scrooge {
+  type Parquet346ScroogeScheme[T <: com.twitter.scrooge.ThriftStruct] = com.twitter.scalding.parquet.cascading.scrooge.Parquet346ScroogeScheme[T]
+}
+
