+++ scalding-parquet/src/main/scala/com/twitter/scalding/parquet/package.scala
@@ -1,0 +1,10 @@
+package com.twitter.scalding
+
+package object parquet {
+  type ParquetValueScheme[T] = com.twitter.scalding.parquet.cascading.ParquetValueScheme[T]
+
+  object ParquetValueScheme {
+    type Config[T] = com.twitter.scalding.parquet.cascading.ParquetValueScheme.Config[T]
+  }
+}
+
