+++ scalding-parquet/src/main/scala/com/twitter/scalding/parquet/thrift/package.scala
@@ -1,0 +1,6 @@
+package com.twitter.scalding.parquet
+
+package object thrift {
+  type Parquet346TBaseScheme[T <: org.apache.thrift.TBase[_, _]] = com.twitter.scalding.parquet.cascading.thrift.Parquet346TBaseScheme[T]
+}
+
