+++ scalding-parquet/src/main/scala/com/twitter/scalding/parquet/tuple/package.scala
@@ -1,0 +1,6 @@
+package com.twitter.scalding.parquet
+
+package object tuple {
+  type ParquetTupleScheme = com.twitter.scalding.parquet.cascading.tuple.ParquetTupleScheme
+}
+
